Create mysql.php
mysql.php - for include in change to views/listing.php to grab current show from MythTV's line up. 

I know I need to write an IF - $User MythTV Listings = True the section in listing.php  But it's a start and I will continue to play with it. 

Please post feed back if you have any.